move VaultHistoryEntry to separated file
